requirements.txt
tqdm
numpy
opencv-python
Pillow
scipy
bezier
matplotlib
tensorboardX
